+++ src/verifier/options.rs
@@ -5,10 +5,20 @@
 use crate::backend::{BackendHook, NoHook};
 
 /// The optional arguments of a verification.
-#[derive(Debug, Clone)]
+#[derive(Debug)]
 pub struct VerifyOptions<BH: BackendHook = NoHook> {
     hook_init_arg: BH::InitArg,
 }
+
+// Hand-rolled: a derive would require the hook itself to be [Clone]/[Copy], when the only thing
+// held is its init argument, which [BackendHook] already requires to be [Copy].
+impl<BH: BackendHook> Clone for VerifyOptions<BH> {
+    fn clone(&self) -> Self {
+        return *self;
+    }
+}
+
+impl<BH: BackendHook> Copy for VerifyOptions<BH> {}
 
 impl VerifyOptions<NoHook> {
     /// Creates the default options.